renamed project from slack2bm to slackaps
