+++ MANIFEST.in
@@ -1,1 +1,1 @@
-include slack2bm/data/*
+include slackaps/data/*
