--- bin/slack2bm
@@ -1,82 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-
-"""
-.. module:: metah5.py
-   :platform: Unix
-   :synopsis: Reads the meta data of a tomographic hdf5 file and generates a table (rst) compatible with sphynx/readthedocs.
-
-"""
-
-import os
-import time
-import pathlib
-import json
-
-from dotenv import load_dotenv
-from datetime import datetime
-from slack_bolt import App 
-from slack_bolt.adapter.socket_mode import SocketModeHandler 
-from pathlib import Path
-
-from slack2bm import pv
-from slack2bm import log
-
-data_path = Path(__file__).parent / '../../slack2bm/data'
-
-print(data_path)
-# Set bot tokens as environment values
-env_path = os.path.join(str(pathlib.Path.home()), '.slackenv')
-load_dotenv(dotenv_path=env_path)
-
-bot_token = os.environ.get("BOT_TOKEN")
-app_token = os.environ.get("APP_TOKEN") 
-
-app = App(token = bot_token)
-
-@app.event("message") 
-def reply(payload, say, client): 
-    try: 
-        text = payload["text"].lower()
-    except: 
-        return 
-    user = client.users_info(user=payload["user"])["user"]["real_name"] 
-    chan = client.conversations_info(channel=payload["channel"])["channel"]["name"]
-
-    # Load PV lookup table from the JSON file
-    with open(os.path.join(data_path, 'pv.json')) as json_file:
-        f_lookup = json.load(json_file)
-    pvs = {}
-    if chan == "auto_2bm": 
-        if text in f_lookup:
-            pvs = f_lookup[text]
-            time.sleep(1)
-            _, slack_messages = pv.check_pvs_connected(pvs)
-            for message in slack_messages:
-                say(message)
-        elif text == 'help':
-            for text in f_lookup:
-                say(text)
-    else:
-        if text_match(text, "hello"):
-            say("Greetings *{0}*!".format(user)) 
-
-def start_robot():
-
-    SocketModeHandler(app, app_token).start() 
-
-def main():
-    # set logs directory
-    home = os.path.expanduser("~")
-    logs_home = home + '/logs/'
-    # make sure logs directory exists
-    if not os.path.exists(logs_home):
-        os.makedirs(logs_home)
-    # setup logger
-    lfname = logs_home + 'slack_' + datetime.strftime(datetime.now(), "%Y-%m-%d_%H:%M:%S") + '.log'
-    log.setup_custom_logger(lfname)
-
-    start_robot()
-
-if __name__ == '__main__':
-    main()
