+++ setup.py
@@ -4,15 +4,15 @@ import os
 
 
 setup(
-    name='slack2bm',
+    name='slackaps',
     version=open('VERSION').read().strip(),
     #version=__version__,
     author='Francesco De Carlo, Tao Zhou, Fanny Rodolakis, Byeongdu Lee',
     author_email='decarlof@gmail.com',
-    url='https://github.com/xray-imaging/slack2bm',
+    url='https://github.com/xray-imaging/slackaps',
     packages=find_packages(),
     include_package_data = True,
-    scripts=['bin/slack2bm'],
-    description='cli to run a slack rob at 2-bm',
+    scripts=['bin/slackaps'],
+    description='cli to run a slack rob at the APS',
     zip_safe=False,
 )
