--- slack2bm/data/pv.json
@@ -1,1 +1,0 @@
-{"ring": {"s_current": "S:SRcurrentAI", "shutter_status": "PA:02BM:STA_A_FES_OPEN_PL", "acis_shutter_permit": "ACIS:ShutterPermit", "s_desired_mode": "S:DesiredMode", "ops_message1": "OPS:message5", "ops_message2": "OPS:message5", "ops_message3": "OPS:message5", "ops_message4": "OPS:message5", "ops_message5": "OPS:message5"}, "eps": {"fault_temp_m1": "2bm:Fault:Temp:M1", "fault_helium_flow": "2bm:Fault:Helium_Flow", "fault_water_p6_shutter": "2bm:Fault:Water:P6_Shutter", "fault_water_dmm": "2bm:Fault:Water:DMM", "fault_water_bma_Window": "2bm:Fault:Water:BMA_Window", "fault_water_m1": "2bm:Fault:Water:M1", "fault_water_filter_slits": "2bm:Fault:Water:FilterSlits", "vacuum_bmb_slits": "2bm:Vacuum:BMB_Slits", "vacuum_mini_hutch_2": "2bm:Vacuum:mini_hutch_2", "vacuum_mini_hutch_1": "2bm:Vacuum:mini_hutch_1", "vacuum_p6_shutter": "2bm:Vacuum:P6_Shutter", "vacuum_dmm": "2bm:Vacuum:DMM", "vacuum_m1": "2bm:Vacuum:M1", "vacuum_filter_slits": "2bm:Vacuum:FilterSlits"}, "energy": {"energy": "2bma:TomoScan:Energy", "energy_mode": "2bma:TomoScan:EnergyMode", "filters": "2bma:TomoScan:Filters"}, "optics": {"scintillator_type": "2bma:TomoScan:ScintillatorType", "scintillator_thickness": "2bma:TomoScan:ScintillatorThickness", "image_pixel_size": "2bma:TomoScan:ImagePixelSize", "detector_pixel_size": "2bma:TomoScan:DetectorPixelSize", "camera_objective": "2bma:TomoScan:cam_eraObjective", "camera_tube_lens": "2bma:TomoScan:cam_eraTubeLength"}, "sample": {"sameple_name": "2bma:TomoScan:SampleName", "sample_description_1": "2bma:TomoScan:SampleDescription1", "sample_description_2": "2bma:TomoScan:SampleDescription2", "sample_description_3": "2bma:TomoScan:SampleDescription3"}, "user": {"user_name": "2bma:TomoScan:UserName", "user_last_name": "2bma:TomoScan:UserLastName", "user_institution": "2bma:TomoScan:UserInstitution", "user_badge": "2bma:TomoScan:UserBadge", "user_email": "2bma:TomoScan:UserEmail", "proposal_number": "2bma:TomoScan:ProposalNumber", "proposal_title": "2bma:TomoScan:ProposalTitle", "esaf_number": "2bma:TomoScan:ESAFNumber", "user_info_update": "2bma:TomoScan:UserInfoUpdate"}, "data": {"detector_top_dir": "2bma:TomoScan:DetectorTopDir", "user_last_name": "2bma:TomoScan:UserLastName", "experiment_year_month": "2bma:TomoScan:ExperimentYearMonth", "remote_data_analysis_dir": "2bma:TomoScan:RemoteAnalysisDir", "copy_to_analysis_dir": "2bma:TomoScan:CopyToAnalysisDir"}, "scan": {"testing": "2bma:TomoScan:Testing", "rotation_start": "2bma:TomoScan:RotationStart\"", "rotation_step": "2bma:TomoScan:RotationStep", "rotation_stop": "2bma:TomoScan:RotationStop", "num_angle": "2bma:TomoScan:NumAngles", "return_rotation": "2bma:TomoScan:ReturnRotation", "num_dark_fields": "2bma:TomoScan:NumDarkFields", "dark_field_mode": "2bma:TomoScan:DarkFieldMode", "dark_field_value": "2bma:TomoScan:DarkFieldValue", "num_flat_fields": "2bma:TomoScan:NumFlatFields", "flat_field_mode": "2bma:TomoScan:FlatFieldMode", "flat_field_axis": "2bma:TomoScan:FlatFieldAxis", "flat_field_value": "2bma:TomoScan:FlatFieldValue", "sample_in_x": "2bma:TomoScan:SampleInX", "sample_out_x": "2bma:TomoScan:SampleOutX", "sample_in_y": "2bma:TomoScan:SampleInY", "sample_out_y": "2bma:TomoScan:SampleOutY", "move_sample_in": "2bma:TomoScan:MoveSampleIn", "move_sample_out": "2bma:TomoScan:MoveSampleOut", "start_scan": "2bma:TomoScan:StartScan", "abort_scan": "2bma:TomoScan:AbortScan", "scan_status": "2bma:TomoScan:ScanStatus", "images_collected": "2bma:TomoScan:ImagesCollected", "images_saved": "2bma:TomoScan:ImagesSaved", "elapsed_time": "2bma:TomoScan:ElapsedTime", "remaining_time": "2bma:TomoScan:RemainingTime", "server_running": "2bma:TomoScan:ServerRunning", "frame_type": "2bma:TomoScan:FrameType"}, "file": {"file_path": "2bma:TomoScan:FilePath", "file_name": "2bma:TomoScan:FileName", "file_path_exists": "2bma:TomoScan:FilePathExists", "overwrite_warning": "2bma:TomoScan:OverwriteWarning", "hdf5_projection_location": "2bma:TomoScan:HDF5ProjectionLocation", "hdf5_dark_location": "2bma:TomoScan:HDF5DarkLocation", "hdf5_flat_location": "2bma:TomoScan:HDF5FlatLocation", "hdf5_location": "2bma:TomoScan:HDF5Location", "FPNumCaptured": "2bmbSP1:HDF1:NumCaptured_RBV", "FPFullFileName": "2bmbSP1:HDF1:FullFileName_RBV"}, "detector": {"exposure_time": "2bma:TomoScan:ExposureTime", "flat_field_time": "2bma:TomoScan:FlatExposureTime", "different_flat_exposure": "2bma:TomoScan:DifferentFlatExposure", "cam_manufacturer": "2bmbSP1:cam1:Manufacturer_RBV", "cam_model": "2bmbSP1:cam1:Model_RBV", "cam_acquire": "2bmbSP1:cam1:Acquire", "cam_acquire_time": "2bmbSP1:cam1:AcquireTime", "cam_acquire_busy": "2bmbSP1:cam1:AcquireBusy", "cam_image_mode": "2bmbSP1:cam1:ImageMode", "cam_trigger_mode": "2bmbSP1:cam1:TriggerMode", "cam_num_images": "2bmbSP1:cam1:NumImages", "cam_num_images_counter": "2bmbSP1:cam1:NumImagesCounter_RBV"}}
